code/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.8.3=pypi_0
aiosignal=1.3.1=pypi_0
astroid=2.12.13=py310h5588dad_0
asttokens=2.1.0=pyhd8ed1ab_0
async-timeout=4.0.2=pypi_0
attrs=22.1.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=2.116=mkl
blas-devel=3.9.0=16_win64_mkl
boto3=1.28.80=pypi_0
botocore=1.31.80=pypi_0
brotlipy=0.7.0=py310h8d17308_1005
bzip2=1.0.8=h8ffe710_4
ca-certificates=2022.12.7=h5b45459_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py310h628cb3f_2
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=win_pyhd8ed1ab_2
cloudpickle=2.2.1=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.0.7=pypi_0
cryptography=38.0.3=py310h6e82f81_0
cuda=11.6.2=0
cuda-cccl=11.6.55=0
cuda-command-line-tools=11.6.2=0
cuda-compiler=11.6.2=0
cuda-cudart=11.6.55=0
cuda-cudart-dev=11.6.55=0
cuda-cuobjdump=11.6.124=0
cuda-cupti=11.6.124=0
cuda-cuxxfilt=11.6.124=0
cuda-libraries=11.6.2=0
cuda-libraries-dev=11.6.2=0
cuda-memcheck=11.8.86=0
cuda-nsight-compute=11.8.0=0
cuda-nvcc=11.6.124=0
cuda-nvdisasm=11.8.86=0
cuda-nvml-dev=11.6.55=0
cuda-nvprof=11.8.87=0
cuda-nvprune=11.6.124=0
cuda-nvrtc=11.6.124=0
cuda-nvrtc-dev=11.6.124=0
cuda-nvtx=11.6.124=0
cuda-nvvp=11.8.87=0
cuda-runtime=11.6.2=0
cuda-sanitizer-api=11.8.86=0
cuda-toolkit=11.6.2=0
cuda-tools=11.6.2=0
cuda-visual-tools=11.6.2=0
cycler=0.11.0=pypi_0
dataclasses=0.8=pyhc8e2a94_3
datasets=2.7.1=pypi_0
debugpy=1.6.4=py310h00ffb61_0
decorator=5.1.1=pyhd8ed1ab_0
dill=0.3.6=pyhd8ed1ab_1
entrypoints=0.4=pyhd8ed1ab_0
evaluate=0.4.0=pypi_0
executing=1.2.0=pyhd8ed1ab_0
filelock=3.8.0=pyhd8ed1ab_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h546665d_0
frozenlist=1.3.3=pypi_0
fsspec=2022.11.0=pypi_0
huggingface_hub=0.11.0=py_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=5.0.0=pyha770c72_1
importlib_metadata=5.0.0=hd8ed1ab_1
intel-openmp=2022.1.0=h57928b3_3787
ipykernel=6.17.1=pyh025b116_0
ipython=8.6.0=pyh08f2357_1
isort=5.10.1=pyhd8ed1ab_0
jedi=0.18.1=pyhd8ed1ab_2
jmespath=1.0.1=pypi_0
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=h8ffe710_2
jupyter_client=7.4.8=pyhd8ed1ab_0
jupyter_core=5.1.0=py310h5588dad_0
kiwisolver=1.4.4=pypi_0
lazy-object-proxy=1.8.0=py310h8d17308_0
lcms2=2.14=h90d422f_0
lerc=4.0.0=h63175ca_0
libblas=3.9.0=16_win64_mkl
libcblas=3.9.0=16_win64_mkl
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.0.86=0
libcurand-dev=10.3.0.86=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.14=hcfcfb64_0
libffi=3.4.2=h8ffe710_5
liblapack=3.9.0=16_win64_mkl
liblapacke=3.9.0=16_win64_mkl
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.38=h19919ed_0
libprotobuf=3.21.9=h12be248_0
libsodium=1.0.18=h8d14728_1
libsqlite=3.40.0=hcfcfb64_0
libtiff=4.4.0=h8e97e67_4
libuv=1.44.2=h8ffe710_0
libwebp-base=1.2.4=h8ffe710_0
libxcb=1.13=hcd874cb_1004
libzlib=1.2.13=hcfcfb64_4
lightgbm=3.3.4=pypi_0
llvmlite=0.39.1=py310hb84602e_1
loguru=0.6.0=py310h5588dad_2
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
matplotlib=3.6.3=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mccabe=0.7.0=pyhd8ed1ab_0
mkl=2022.1.0=h6a75c08_874
mkl-devel=2022.1.0=h57928b3_875
mkl-include=2022.1.0=h6a75c08_874
msys2-conda-epoch=20160418=1
multidict=6.0.3=pypi_0
multiprocess=0.70.14=pypi_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
nsight-compute=2022.3.0.22=0
numba=0.56.4=py310h19bcfe9_0
numpy=1.23.4=py310h4a8f9c9_1
openjpeg=2.5.0=hc9384bd_1
openssl=3.0.8=hcfcfb64_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.5.1=py310h1c4a608_1
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=9.2.0=py310hd4fb230_3
pip=22.3.1=pyhd8ed1ab_0
platformdirs=2.5.2=pyhd8ed1ab_1
prompt-toolkit=3.0.32=pyha770c72_0
protobuf=4.21.9=py310h00ffb61_0
psutil=5.9.4=py310h8d17308_0
pthread-stubs=0.4=hcd874cb_1001
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=10.0.1=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.13.0=pyhd8ed1ab_0
pylint=2.15.6=pyhd8ed1ab_0
pyopenssl=22.1.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=pyh0701188_6
python=3.10.6=hcf16a7b_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytorch=1.13.0=py3.10_cuda11.6_cudnn8_0
pytorch-cuda=11.6=h867d48c_0
pytorch-mutex=1.0=cuda
pytorch-pretrained-bert=0.6.2=pypi_0
pytorchtools=0.0.2=pypi_0
pytz=2022.6=pyhd8ed1ab_0
pywin32=304=py310h00ffb61_2
pyyaml=6.0=py310h8d17308_5
pyzmq=24.0.1=py310hcd737a0_1
regex=2022.10.31=py310h8d17308_0
requests=2.28.1=pyhd8ed1ab_1
responses=0.18.0=pypi_0
s3transfer=0.7.0=pypi_0
sacremoses=master=py_0
scikit-learn=1.1.3=py310had3394f_1
scipy=1.9.3=py310h578b7cb_2
setuptools=65.5.1=pyhd8ed1ab_0
shap=0.41.0=py310hf5e1058_0
six=1.16.0=pyh6c4a22f_0
slicer=0.0.7=pyhd8ed1ab_0
stack_data=0.6.1=pyhd8ed1ab_0
tbb=2021.7.0=h91493d7_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h8ffe710_0
tokenizers=0.11.4=py310he5181cf_1
tomli=2.0.1=pyhd8ed1ab_0
tomlkit=0.11.6=pyha770c72_0
torchaudio=0.13.0=pypi_0
torchvision=0.14.0=pypi_0
tornado=6.2=py310h8d17308_1
tqdm=4.64.1=pyhd8ed1ab_0
traitlets=5.5.0=pyhd8ed1ab_0
transformers=4.24.0=py_0
typing=3.10.0.0=pyhd8ed1ab_0
typing-extensions=4.4.0=hd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022f=h191b570_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=1.26.11=pyhd8ed1ab_0
vc=14.3=h3d8a991_9
vs2015_runtime=14.32.31332=h1d6e394_9
wcwidth=0.2.5=pyh9f0ad1d_2
wheel=0.38.4=pyhd8ed1ab_0
win32_setctime=1.1.0=pyhd8ed1ab_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
wrapt=1.14.1=py310h8d17308_1
xorg-libxau=1.0.9=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xxhash=3.1.0=pypi_0
xz=5.2.6=h8d14728_0
yaml=0.2.5=h8ffe710_2
yarl=1.8.2=pypi_0
zeromq=4.3.4=h0e60522_1
zipp=3.10.0=pyhd8ed1ab_0
zstd=1.5.2=h7755175_4
